+++ frontend/progress.h
@@ -27,7 +27,7 @@ typedef struct {
     uint64_t current_input_samples;
     uint64_t total_input_samples;
     uint32_t sample_rate;
-    uint32_t num_channels;
+    uint16_t num_channels;
     uint32_t current_frame;
     uint32_t total_frames;
     uint64_t total_bytes_written;
